projects/system-overlord-phase0/bots/requirements.txt
firebase-admin==6.5.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
pydantic==2.5.0
